+++ lib/src/yaml/yaml.dart
@@ -11,7 +11,10 @@ import '../serializer/serializer.dart';
  * Same usage as [SerializerRepo]
  */
 class YamlRepo extends SerializerRepo {
-  YamlRepo({List<Serializer> serializers, String typeKey: defaultTypeInfoKey, bool withType = false})
+  YamlRepo(
+      {List<Serializer> serializers,
+      String typeKey: defaultTypeInfoKey,
+      bool withType: false})
       : super(serializers: serializers, typeKey: typeKey, withType: withType);
 
   ///@nodoc