+++ test/serializer/various/various.dart
@@ -142,17 +142,17 @@ class WithIgnoreSerializer extends Serializer<WithIgnore>
   WithIgnore createModel() => new WithIgnore(null, null);
 }
 
-@GenSerializer(fields: const {'original': const EnDecode('renamed')})
+@GenSerializer(fields: const {'original': const EnDecode(alias: 'renamed')})
 class ModelRenamedSerializer extends Serializer<ModelRenamed>
     with _$ModelRenamedSerializer {
   @override
   ModelRenamed createModel() => new ModelRenamed(null);
 }
 
-@GenSerializer(processors: const {
-  'dates': const DateTimeProcessor(),
-  'dynamicMap': const RawData(),
-  'dynamicList': const RawData(),
+@GenSerializer(fields: const {
+  'dates': const Property(processor: const DateTimeProcessor()),
+  'dynamicMap': const Property(processor: const RawData()),
+  'dynamicList': const Property(processor: const RawData()),
 }, serializers: const [
   WithIgnoreSerializer,
 ])